+++ components/camel-spring-rabbitmq/src/test/java/org/apache/camel/component/springrabbit/integration/RabbitMQBasicIT.java
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @Disabled
-public class RabbitMQBasicIntTest extends AbstractRabbitMQIntTest {
+public class RabbitMQBasicIT extends RabbitMQITSupport {
 
     String foo;
     String bar;
