+++ components/camel-spring-rabbitmq/src/test/java/org/apache/camel/component/springrabbit/integration/RabbitMQProducerNullBodyIT.java
@@ -16,23 +16,31 @@
  */
 package org.apache.camel.component.springrabbit.integration;
 
+import org.apache.camel.CamelContext;
 import org.apache.camel.RoutesBuilder;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.springrabbit.SpringRabbitMQComponent;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.amqp.core.AmqpAdmin;
-import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitAdmin;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 
-public class RabbitMQProducerToDIntTest extends AbstractRabbitMQIntTest {
+public class RabbitMQProducerNullBodyIT extends RabbitMQITSupport {
+
+    @Override
+    protected CamelContext createCamelContext() throws Exception {
+        CamelContext camelContext = super.createCamelContext();
+        SpringRabbitMQComponent rmq = camelContext.getComponent("spring-rabbitmq", SpringRabbitMQComponent.class);
+        rmq.setAllowNullBody(true);
+        return camelContext;
+    }
 
     @Test
-    public void testToD() throws Exception {
+    public void testProducer() {
         ConnectionFactory cf = context.getRegistry().lookupByNameAndType("myCF", ConnectionFactory.class);
 
         Queue q = new Queue("myqueue");
@@ -43,23 +51,7 @@ public class RabbitMQProducerToDIntTest extends AbstractRabbitMQIntTest {
         admin.declareExchange(t);
         admin.declareBinding(BindingBuilder.bind(q).to(t).with("foo.bar.#"));
 
-        fluentTemplate.to("direct:start").withBody("Hello World").withHeader("whereTo", "foo").withHeader("myKey", "foo.bar")
-                .send();
-
-        AmqpTemplate template = new RabbitTemplate(cf);
-        String out = (String) template.receiveAndConvert("myqueue");
-        Assertions.assertEquals("Hello World", out);
-
-        fluentTemplate.to("direct:start").withBody("Bye World").withHeader("whereTo", "foo").withHeader("myKey", "foo.bar.baz")
-                .send();
-
-        template = new RabbitTemplate(cf);
-        out = (String) template.receiveAndConvert("myqueue");
-        Assertions.assertEquals("Bye World", out);
-
-        // there should only be 1 rabbit endpoint
-        long count = context.getEndpoints().stream().filter(e -> e.getEndpointUri().startsWith("spring-rabbit")).count();
-        Assertions.assertEquals(1, count);
+        Assertions.assertDoesNotThrow(() -> template.sendBody("direct:start", null));
     }
 
     @Override
@@ -68,7 +60,7 @@ public class RabbitMQProducerToDIntTest extends AbstractRabbitMQIntTest {
             @Override
             public void configure() throws Exception {
                 from("direct:start")
-                        .toD("spring-rabbitmq:${header.whereTo}?routingKey=${header.myKey}");
+                        .to("spring-rabbitmq:foo?routingKey=foo.bar");
             }
         };
     }
